% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_fsom_plots.R
\name{echo.boxplot}
\alias{echo.boxplot}
\title{Plot ECHO-specific boxplot(s)}
\usage{
echo.boxplot(
  echo.melted.mdatframe,
  variable.value,
  x.var = "visit",
  drop.col.var = NULL,
  make.facets = F
)
}
\arguments{
\item{echo.melted.mdatframe}{a 'reshape2::melt()' data.frame; returned from 'echo.melt()'}

\item{variable.value}{cluster/node value (#) to plot}

\item{x.var}{x-axis variable to plot; factor}

\item{drop.col.var}{a paired character 'c(factor,level)'; drops a factor level}

\item{make.facets}{logical; TRUE results in a facet_wrap of clusters}
}
\value{
a ggplot boxplot; a facet_wrapped ggplot boxplot
}
\description{
Plot ECHO-specific boxplot(s)
}
