% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_fsom_plots.R
\name{echo.boxplot.single}
\alias{echo.boxplot.single}
\title{Plot a single ECHO-specific boxplot}
\usage{
echo.boxplot.single(
  echo.melted.mdatframe,
  variable.value,
  x.var = "visit",
  drop.col.var = NULL
)
}
\arguments{
\item{echo.melted.mdatframe}{a 'reshape2::melt()' data.frame; returned from 'echo.melt()'}

\item{variable.value}{cluster/node value (#) to plot}

\item{x.var}{x-axis variable to plot; factor}

\item{drop.col.var}{a paired character 'c(factor,level)'; drops a factor level}
}
\value{
a ggplot boxplot
}
\description{
Plot a single ECHO-specific boxplot
}
