% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_fsom_utility.R
\name{fsom.initialize}
\alias{fsom.initialize}
\title{Build an fsom object using a list of flowFrames}
\usage{
fsom.initialize(fcs.list)
}
\arguments{
\item{fcs.list}{a list of flowFrames (named); returned from 'ECHOfcs::read.fcs.selected.markers'}
}
\value{
fsom object; data columns renamed and transformed; meta-data added
}
\description{
Build an fsom object using a list of flowFrames
}
