% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_fsom_utility.R
\name{generate.clusters}
\alias{generate.clusters}
\title{FlowSOM::metaClustering_consensus modification; generates cluster factor}
\usage{
generate.clusters(dat.mat, max.clusters)
}
\arguments{
\item{dat.mat}{numeric matrix; fsom$map$codes}

\item{max.clusters}{final number of clusters; factored}
}
\value{
numeric factor; length = nrow/# of fsom$map$codes, levels = seq(max.clusters)
}
\description{
FlowSOM::metaClustering_consensus modification; generates cluster factor
}
