% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_fsom_utility.R
\name{generate.counts}
\alias{generate.counts}
\title{Generate per-sample counts (cluster/node)}
\usage{
generate.counts(fsom = fsom)
}
\arguments{
\item{fsom}{fsom (fsom.object in environment)}
}
\value{
a list of length 2; counts$cluster and counts$node
}
\description{
Generate per-sample counts (cluster/node)
}
