% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_fsom_utility.R
\name{map.samples.to.fsom.codes}
\alias{map.samples.to.fsom.codes}
\title{Map samples to existing FlowSOM codes/clusters}
\usage{
map.samples.to.fsom.codes(fcs.file.paths, fsom.codes.path)
}
\arguments{
\item{fcs.file.paths}{character vector of .fcs file paths; returned from 'ECHOfcs::get.fcs.paths.echo()'}

\item{fsom.codes.path}{existing codes/clusters list saved as .rds}
}
\value{
a per-sample list of mapped nodes/clusters; essentially a per-sample,per-event/row node/cluster assignment/index
}
\description{
Map samples to existing FlowSOM codes/clusters
}
