% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_fsom_plots.R
\name{plheat}
\alias{plheat}
\title{Cluster medians heatmap: plotly heatmap with hclust dendrogram ordering}
\usage{
plheat(cluster.medians, break.vec = seq(0, 1, by = 0.05))
}
\arguments{
\item{cluster.medians}{cluster median values; returned from 'generate.cluster.medians()'}

\item{break.vec}{a numeric vector for generating color breaks}
}
\value{
Plotly heatmap
}
\description{
Cluster medians heatmap: plotly heatmap with hclust dendrogram ordering
}
