% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_fsom_plots.R
\name{somnambulate}
\alias{somnambulate}
\title{Interactive Shiny; various plots of FlowSOM results}
\usage{
somnambulate(
  fsom.somnambulated.rds.path = NULL,
  fsom.somnambulated = NULL,
  markers = NULL
)
}
\arguments{
\item{fsom.somnambulated.rds.path}{a file.path to exisiting fsom; returned from 'fsom.somnambulation()'}

\item{fsom.somnambulated}{object returned from 'fsom.somnambulation.mod()'}

\item{markers}{character vector of length 2; marker pairs for initial plots}
}
\value{
launches an interactive window ('shiny::shinyApp()')
}
\description{
Interactive Shiny; various plots of FlowSOM results
}
