% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_fsom_plots.R
\name{plheat}
\alias{plheat}
\title{Cluster medians heatmap: plotly heatmap with hclust dendrogram ordering}
\usage{
plheat(cluster.medians)
}
\arguments{
\item{cluster.medians}{cluster median values; returned from 'generate.cluster.medians()'}
}
\value{
Plotly heatmap
}
\description{
Cluster medians heatmap: plotly heatmap with hclust dendrogram ordering
}
